+++ src/Resources/contao/dca/tl_championslists_categories.php
@@ -1,0 +1,302 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2014 Leo Feyer
+ *
+ * @package News
+ * @link    https://contao.org
+ * @license http://www.gnu.org/licenses/lgpl-3.0.html LGPL
+ */
+
+
+/**
+ * Table tl_championslists_categories
+ */
+$GLOBALS['TL_DCA']['tl_championslists_categories'] = array
+(
+
+	// Config
+	'config' => array
+	(
+		'dataContainer'               => 'Table',
+		'switchToEdit'                => true,
+		'enableVersioning'            => true,
+		'sql' => array
+		(
+			'keys' => array
+			(
+				'id'    => 'primary'
+			)
+		)
+	),
+
+	// List
+	'list' => array
+	(
+		'sorting' => array
+		(
+			'mode'                    => 2,
+			'fields'                  => array('title', 'alias'),
+			'flag'                    => 1,
+			'panelLayout'             => 'filter;sort,search,limit'
+		),
+		'label' => array
+		(
+			'fields'                  => array('title', 'alias'),
+			'showColumns'             => true,
+			'format'                  => '%s'
+		),
+		'global_operations' => array
+		(
+			'listen' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['listen'],
+				'href'                => 'table=tl_championslists',
+				'icon'                => 'bundles/contaochampionslists/images/icon.png',
+				'attributes'          => 'onclick="Backend.getScrollOffset();"'
+			),
+			'all' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['MSC']['all'],
+				'href'                => 'act=select',
+				'class'               => 'header_edit_all',
+				'attributes'          => 'onclick="Backend.getScrollOffset()" accesskey="e"'
+			)
+		),
+		'operations' => array
+		(
+			'edit' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['edit'],
+				'href'                => 'act=edit',
+				'icon'                => 'edit.gif'
+			),
+			'copy' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['copy'],
+				'href'                => 'act=copy',
+				'icon'                => 'copy.gif',
+			),
+			'delete' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['delete'],
+				'href'                => 'act=delete',
+				'icon'                => 'delete.gif',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\'))return false;Backend.getScrollOffset()"',
+			),
+			'toggle' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['toggle'],
+				'icon'                => 'visible.gif',
+				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
+				'button_callback'     => array('tl_championslists_categories', 'toggleIcon')
+			),
+			'show' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_championslists_categories']['show'],
+				'href'                => 'act=show',
+				'icon'                => 'show.gif'
+			)
+		)
+	),
+
+	// Palettes
+	'palettes' => array
+	(
+		'default'                     => '{title_legend},title,alias;{publish_legend},published'
+	),
+
+	// Fields
+	'fields' => array
+	(
+		'id' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL auto_increment"
+		),
+		'sorting' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default 0"
+		),
+		'tstamp' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+		'title' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_championslists_categories']['title'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array
+			(
+				'mandatory'           => true,
+				'maxlength'           => 64,
+				'tl_class'            => 'w50'
+			),
+			'sql'                     => "varchar(64) NOT NULL default ''"
+		),
+		'alias' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_championslists_categories']['alias'],
+			'exclude'                 => true,
+			'search'                  => true,
+			'inputType'               => 'text',
+			'eval'                    => array
+			(
+				'rgxp'                => 'alias',
+				'unique'              => true,
+				'maxlength'           => 64,
+				'doNotCopy'           => true,
+				'tl_class'            => 'w50'
+			),
+			'save_callback' => array
+			(
+				array('tl_championslists_categories', 'generateAlias')
+			),
+			'sql'                     => "varbinary(64) NOT NULL default ''"
+		), 
+		'published' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_championslists_categories']['published'],
+			'exclude'                 => true,
+			'filter'                  => true,
+			'flag'                    => 1,
+			'default'                 => true,
+			'inputType'               => 'checkbox',
+			'eval'                    => array
+			(
+				'doNotCopy'           => true
+			),
+			'sql'                     => "char(1) NOT NULL default ''"
+		),  
+	)
+);
+
+
+/**
+ * Class tl_championslists_categories
+ *
+ * Provide miscellaneous methods that are used by the data configuration array.
+ * @copyright  Leo Feyer 2005-2014
+ * @author     Leo Feyer <https://contao.org>
+ * @package    News
+ */
+class tl_championslists_categories extends Backend
+{
+
+	/**
+	 * Import the back end user object
+	 */
+	public function __construct()
+	{
+		parent::__construct();
+		$this->import('BackendUser', 'User');
+	}
+
+	/**
+	 * Generiert automatisch ein Alias aus dem Titel
+	 * @param mixed
+	 * @param \DataContainer
+	 * @return string
+	 * @throws \Exception
+	 */
+	public function generateAlias($varValue, DataContainer $dc)
+	{
+		$autoAlias = false;
+
+		// Generate alias if there is none
+		if($varValue == '')
+		{
+			$autoAlias = true;
+			$varValue = str_replace('id-', '', \StringUtil::standardize(\StringUtil::restoreBasicEntities($dc->activeRecord->title)));
+		}
+
+		// Check whether the news alias exists
+		if($varValue == 'meister')
+		{
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['tl_championslists_categories']['error_alias'], $varValue));
+		}
+
+		$objAlias = $this->Database->prepare("SELECT id FROM tl_championslists_categories WHERE alias=?")
+		                           ->execute($varValue);
+
+		// Check whether the news alias exists
+		if($objAlias->numRows > 1 && !$autoAlias)
+		{
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['aliasExists'], $varValue));
+		}
+
+		// Add ID to alias
+		if($objAlias->numRows && $autoAlias)
+		{
+			$varValue .= '-' . $dc->id;
+		}
+
+		return $varValue;
+	} 
+
+	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
+	{
+		$this->import('BackendUser', 'User');
+
+		if (strlen($this->Input->get('tid')))
+		{
+			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
+			$this->redirect($this->getReferer());
+		}
+
+		// Check permissions AFTER checking the tid, so hacking attempts are logged
+		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_championslists_categories::published', 'alexf'))
+		{
+			return '';
+		}
+
+		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
+
+		if (!$row['published'])
+		{
+			$icon = 'invisible.gif';
+		}
+
+		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
+	}
+
+	public function toggleVisibility($intId, $blnPublished)
+	{
+		// Check permissions to publish
+		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_championslists_categories::published', 'alexf'))
+		{
+			$this->log('Kein Zugriffsrecht für Aktivierung Datensatz ID "'.$intId.'"', 'tl_championslists_categories toggleVisibility', TL_ERROR);
+			// Zurücklink generieren, ab C4 ist das ein symbolischer Link zu "contao"
+			if (version_compare(VERSION, '4.0', '>='))
+			{
+				$backlink = \System::getContainer()->get('router')->generate('contao_backend');
+			}
+			else
+			{
+				$backlink = 'contao/main.php';
+			}
+			$this->redirect($backlink.'?act=error');
+		}
+		
+		$this->createInitialVersion('tl_championslists_categories', $intId);
+		
+		// Trigger the save_callback
+		if (is_array($GLOBALS['TL_DCA']['tl_championslists_categories']['fields']['published']['save_callback']))
+		{
+			foreach ($GLOBALS['TL_DCA']['tl_championslists_categories']['fields']['published']['save_callback'] as $callback)
+			{
+				$this->import($callback[0]);
+				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
+			}
+		}
+		
+		// Update the database
+		$this->Database->prepare("UPDATE tl_championslists_categories SET tstamp=". time() .", published='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
+					   ->execute($intId);
+		$this->createNewVersion('tl_championslists_categories', $intId);
+	}
+
+}