+++ src/Resources/contao/languages/de/default.php
@@ -1,4 +1,6 @@
 <?php
 
-$GLOBALS['TL_LANG']['CTE']['championslists'] = array('Meisterliste', 'Eine Meisterliste einfügen');
 $GLOBALS['TL_LANG']['CTE']['champion'] = array('Aktueller Meister', 'Eine aktuellen Meister aus einer Meisterliste einfügen');
+
+$GLOBALS['TL_LANG']['CTE']['championslists_mono'] = array('Meisterliste Einzelwettbewerb', 'Bitte wählen Sie die zu verwendende Meisterliste Einzelwetttbewerb aus.');
+$GLOBALS['TL_LANG']['CTE']['championslists_multi'] = array('Meisterliste Mannschaftswettbewerb', 'Bitte wählen Sie die zu verwendende Meisterliste Mannschaftswetttbewerb aus.');